requirements.txt
Copyopenai>=1.0.0
agents>=0.2.0  # OpenAI Agents SDK
pyyaml>=6.0
pydantic>=2.0.0
requests>=2.25.1
gradio>=4.0.0
python-dotenv>=0.19.0
tavily-python>=0.5.1 